requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements.in
#
alembic==1.16.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
arq==0.26.3
    # via -r requirements.in
async-timeout==5.0.1
    # via redis
bcrypt==4.3.0
    # via passlib
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via arq
cohere==5.17.0
    # via -r requirements.in
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==45.0.6
    # via python-jose
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.116.1
    # via -r requirements.in
fastavro==1.12.0
    # via cohere
filelock==3.19.1
    # via huggingface-hub
fsspec==2025.7.0
    # via huggingface-hub
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.3.0
    # via hypercorn
hiredis==3.2.1
    # via redis
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   cohere
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.34.4
    # via tokenizers
hypercorn==0.17.3
    # via -r requirements.in
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via -r requirements.in
langchain-core==0.3.74
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.31
    # via -r requirements.in
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.17
    # via
    #   langchain
    #   langchain-core
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
openai==1.101.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.11.2
    # via langsmith
packaging==25.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pinecone-client==6.0.0
    # via -r requirements.in
pinecone-plugin-interface==0.0.7
    # via pinecone-client
priority==2.0.0
    # via hypercorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.11.7
    # via
    #   -r requirements.in
    #   cohere
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via
    #   cohere
    #   pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pyjwt==2.10.1
    # via redis
python-dateutil==2.9.0.post0
    # via pinecone-client
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
redis[hiredis]==5.3.1
    # via
    #   -r requirements.in
    #   arq
regex==2025.7.34
    # via tiktoken
requests==2.32.5
    # via
    #   -r requirements.in
    #   cohere
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
    #   langchain
starlette==0.47.3
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.11.0
    # via langchain-openai
tokenizers==0.21.4
    # via cohere
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
types-requests==2.32.4.20250809
    # via cohere
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   cohere
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   pinecone-client
    #   requests
    #   types-requests
wsproto==1.2.0
    # via hypercorn
zstandard==0.24.0
    # via langsmith
